refactor: ArrayState 타입 위치를 사용처 모듈로 이관
