+++ src/components/CourseSelection/CourseSelectionView.tsx
@@ -7,10 +7,9 @@ import GradeChip from '@/components/GradeChip';
 import ViewSelectedCoursesButton from '@/components/ViewSelectedCoursesButton';
 import { CourseTypeContext } from '@/context/CourseTypeContext';
 import { courseSelectionInfo, gradeSelection } from '@/data/courseSelectionInfo';
-import { useGetArrayState } from '@/hooks/useGetArrayState';
+import { ArrayState, useGetArrayState } from '@/hooks/useGetArrayState';
 import { Course } from '@/schemas/courseSchema';
 import { Grade } from '@/schemas/studentSchema';
-import { ArrayState } from '@/types/common';
 import { isSameCourse } from '@/utils/course';
 
 interface CourseSelectionSkeletonProps {
