+++ src/hooks/useGetArrayState.ts
@@ -1,6 +1,6 @@
 import { useMemo } from 'react';
 
-import { ArrayState } from '@/types/common';
+export type ArrayState = 'EMPTY' | 'FILLED';
 
 export const useGetArrayState = <T>(array: T[]) => {
   return useMemo((): ArrayState => (array.length === 0 ? 'EMPTY' : 'FILLED'), [array]);
