--- src/types/common.ts
@@ -1,1 +1,0 @@
-export type ArrayState = 'EMPTY' | 'FILLED';
